refactor(password-validation): extract password validation rules to custom hook
