+++ apps/web/app/components/PasswordValidation/PasswordValidationDisplay.tsx
@@ -1,7 +1,6 @@
 import { useFormContext } from "react-hook-form";
-import { useTranslation } from "react-i18next";
 
-import { passwordValidationRules } from "~/modules/Auth/constants";
+import { usePasswordValidationRules } from "~/hooks/usePasswordValidationRules";
 
 import { validatePasswordStrength } from "../../modules/Dashboard/Settings/schema/password.schema";
 import PasswordStrengthBars from "../PasswordStrengthBars";
@@ -13,44 +12,16 @@ interface PasswordValidationDisplayProps {
 }
 
 export default function PasswordValidationDisplay({ fieldName }: PasswordValidationDisplayProps) {
-  const { t } = useTranslation();
   const { watch } = useFormContext();
+  const validationRulesConfig = usePasswordValidationRules();
 
   const password = watch(fieldName) || "";
   const validationResult = validatePasswordStrength(password);
 
-  const validationRulesConfig = [
-    {
-      key: "minLength",
-      getText: () =>
-        t("passwordValidationDisplay.passwordMinLength", {
-          count: passwordValidationRules.minLength,
-        }),
-    },
-    {
-      key: "hasLowerCase",
-      getText: () => t("passwordValidationDisplay.passwordLowercase"),
-    },
-    {
-      key: "hasUpperCase",
-      getText: () => t("passwordValidationDisplay.passwordUppercase"),
-    },
-    {
-      key: "hasNumber",
-      getText: () => t("passwordValidationDisplay.passwordNumber"),
-    },
-    {
-      key: "hasSpecialChar",
-      getText: () => t("passwordValidationDisplay.passwordSpecial"),
-    },
-  ];
-
   return (
     <div id="password-hints" className="mb-3">
       <PasswordStrengthBars fieldName={fieldName} />
-      <h4 className="my-2 text-sm font-semibold text-gray-800">
-        {t("passwordValidationDisplay.header")}
-      </h4>
+      <h4 className="my-2 text-sm font-semibold text-gray-800">Password requirements</h4>
       <ul className="space-y-1 text-sm text-gray-500">
         {validationRulesConfig.map((rule) => (
           <PasswordValidationRuleItem
