+++ apps/web/app/components/PasswordValidation/PasswordValidationRuleItem.tsx
@@ -10,7 +10,11 @@ interface PasswordValidationRuleItemProps {
 export const PasswordValidationRuleItem = ({ isValid, text }: PasswordValidationRuleItemProps) => {
   const Icon = isValid ? CheckIcon : XIcon;
   return (
-    <li className={cn("flex items-center gap-x-2", isValid && "text-teal-500")}>
+    <li
+      className={cn("flex items-center gap-x-2", {
+        "text-teal-500": isValid,
+      })}
+    >
       <Icon className="size-4 shrink-0" />
       {text}
     </li>
